+++ modern/_1.14.4/src/main/kotlin/cn/enaium/noexpensive/command/compatibility.kt
@@ -6,14 +6,17 @@ import cn.enaium.noexpensive.enums.Action
 import com.mojang.brigadier.Command
 import com.mojang.brigadier.CommandDispatcher
 import com.mojang.brigadier.context.CommandContext
-import net.minecraft.command.argument.EnchantmentArgumentType
-import net.minecraft.item.Items
+import net.minecraft.command.arguments.ItemEnchantmentArgumentType
 import net.minecraft.server.command.CommandManager
 import net.minecraft.server.command.ServerCommandSource
-import net.minecraft.text.*
+import net.minecraft.text.HoverEvent
+import net.minecraft.text.LiteralText
+import net.minecraft.text.Text
+import net.minecraft.text.TranslatableText
 import net.minecraft.util.Formatting
 import net.minecraft.util.Identifier
 import net.minecraft.util.registry.Registry
+
 
 /**
  * @author Enaium
@@ -28,41 +31,35 @@ fun compatibilityCommand(dispatcher: CommandDispatcher<ServerCommandSource>) {
                         .then(
                             CommandManager.argument(
                                 "enchantment1",
-                                EnchantmentArgumentType.enchantment()
+                                ItemEnchantmentArgumentType.itemEnchantment()
                             )
                                 .then(
                                     CommandManager.argument(
                                         "enchantment2",
-                                        EnchantmentArgumentType.enchantment()
+                                        ItemEnchantmentArgumentType.itemEnchantment()
                                     ).executes { context: CommandContext<ServerCommandSource> ->
                                         val enchantment1 =
-                                            EnchantmentArgumentType.getEnchantment(context, "enchantment1")
+                                            ItemEnchantmentArgumentType.getEnchantment(context, "enchantment1")
                                         val enchantment2 =
-                                            EnchantmentArgumentType.getEnchantment(context, "enchantment2")
+                                            ItemEnchantmentArgumentType.getEnchantment(context, "enchantment2")
                                         val enchantment1Name = Registry.ENCHANTMENT.getId(enchantment1)!!.toString()
                                         val enchantment2Name = Registry.ENCHANTMENT.getId(enchantment2)!!.toString()
-                                        val enchantment1ItemStack = Items.ENCHANTED_BOOK.defaultStack
-                                        val enchantment2ItemStack = Items.ENCHANTED_BOOK.defaultStack
-                                        enchantment1ItemStack.addEnchantment(enchantment1, 1)
-                                        enchantment2ItemStack.addEnchantment(enchantment2, 1)
-                                        val enchantment1Text =
-                                            TranslatableText(enchantment1Name).styled { style: Style ->
-                                                style.withHoverEvent(
-                                                    HoverEvent(
-                                                        HoverEvent.Action.SHOW_ITEM,
-                                                        HoverEvent.ItemStackContent(enchantment1ItemStack)
-                                                    )
-                                                ).withColor(Formatting.AQUA)
-                                            }
-                                        val enchantment2Text =
-                                            TranslatableText(enchantment2Name).styled { style: Style ->
-                                                style.withHoverEvent(
-                                                    HoverEvent(
-                                                        HoverEvent.Action.SHOW_ITEM,
-                                                        HoverEvent.ItemStackContent(enchantment2ItemStack)
-                                                    )
-                                                ).withColor(Formatting.AQUA)
-                                            }
+                                        val enchantment1Text = LiteralText(enchantment1Name).styled { style ->
+                                            style.setHoverEvent(
+                                                HoverEvent(
+                                                    HoverEvent.Action.SHOW_TEXT,
+                                                    TranslatableText(enchantment1.translationKey)
+                                                )
+                                            ).setColor(Formatting.AQUA)
+                                        }
+                                        val enchantment2Text = LiteralText(enchantment2Name).styled { style ->
+                                            style.setHoverEvent(
+                                                HoverEvent(
+                                                    HoverEvent.Action.SHOW_TEXT,
+                                                    TranslatableText(enchantment2.translationKey)
+                                                )
+                                            ).setColor(Formatting.AQUA)
+                                        }
                                         val compatibility = Config.model.compatibility
                                         if (action === Action.PUT) {
                                             if (compatibility.containsKey(enchantment1Name)) {
@@ -115,50 +112,46 @@ fun compatibilityCommand(dispatcher: CommandDispatcher<ServerCommandSource>) {
             COMPATIBILITY.then(
                 CommandManager.literal("list").executes { context: CommandContext<ServerCommandSource> ->
                     val compatibility = Config.model.compatibility
-                    var previous: MutableText? = null
+                    var previous: Text? = null
                     for ((key, value) in compatibility) {
-                        val keyItemStack = Items.ENCHANTED_BOOK.defaultStack
-                        keyItemStack.addEnchantment(Registry.ENCHANTMENT[Identifier(key)] ?: continue, 1)
-                        val enchantment = LiteralText(key).styled { style: Style ->
-                            style.withHoverEvent(
+                        val enchantment = LiteralText(key).styled { style ->
+                            style.setHoverEvent(
                                 HoverEvent(
-                                    HoverEvent.Action.SHOW_ITEM,
-                                    HoverEvent.ItemStackContent(keyItemStack)
+                                    HoverEvent.Action.SHOW_TEXT,
+                                    TranslatableText(
+                                        Registry.ENCHANTMENT[Identifier(key)]?.translationKey ?: return@styled
+                                    )
                                 )
-                            ).withColor(
-                                Formatting.AQUA
-                            )
+                            ).setColor(Formatting.AQUA)
                         }
                         if (value.isNotEmpty()) {
-                            enchantment.append(
-                                LiteralText(" -> ").styled { style: Style -> style.withColor(Formatting.YELLOW) })
+                            enchantment.append(LiteralText(" -> ").styled { style -> style.setColor(Formatting.YELLOW) })
                             for (s in value) {
-                                val valueItemStack = Items.ENCHANTED_BOOK.defaultStack
-                                valueItemStack.addEnchantment(Registry.ENCHANTMENT[Identifier(key)] ?: continue, 1)
-                                enchantment.append(LiteralText(s).styled { style: Style ->
-                                    style.withHoverEvent(
+                                enchantment.append(LiteralText(s).styled { style ->
+                                    style.setHoverEvent(
                                         HoverEvent(
-                                            HoverEvent.Action.SHOW_ITEM, HoverEvent.ItemStackContent(valueItemStack)
+                                            HoverEvent.Action.SHOW_TEXT,
+                                            TranslatableText(
+                                                Registry.ENCHANTMENT[Identifier(key)]?.translationKey ?: return@styled
+                                            )
                                         )
-                                    ).withColor(Formatting.AQUA)
+                                    ).setColor(Formatting.AQUA)
                                 })
                                 if (s != value[value.size - 1]) {
-                                    enchantment.append(
-                                        LiteralText(", ")
-                                            .styled { style: Style -> style.withColor(Formatting.YELLOW) })
+                                    enchantment.append(LiteralText(", ").styled { style -> style.setColor(Formatting.YELLOW) })
                                 }
                             }
                         }
                         if (previous == null) {
                             previous = enchantment
                         } else {
-                            previous.append(
-                                LiteralText(", ").styled { style: Style -> style.withColor(Formatting.RED) })
+                            previous.append(LiteralText(", ").styled { style -> style.setColor(Formatting.RED) })
                             previous.append(enchantment)
                         }
                     }
-                    val finalPrevious = previous
-                    context.source.sendFeedback(finalPrevious, false)
+                    if (previous != null) {
+                        context.source.sendFeedback(previous, false)
+                    }
                     Command.SINGLE_SUCCESS
                 })
         )