+++ modern/_1.16.5/src/main/kotlin/cn/enaium/noexpensive/command/compatibility.kt
@@ -157,8 +157,9 @@ fun compatibilityCommand(dispatcher: CommandDispatcher<ServerCommandSource>) {
                             previous.append(enchantment)
                         }
                     }
-                    val finalPrevious = previous
-                    context.source.sendFeedback(finalPrevious, false)
+                    if (previous != null) {
+                        context.source.sendFeedback(previous, false)
+                    }
                     Command.SINGLE_SUCCESS
                 })
         )