+++ modern/_1.20.1/src/main/kotlin/cn/enaium/noexpensive/command/compatibility.kt
@@ -170,8 +170,10 @@ fun compatibilityCommand(dispatcher: CommandDispatcher<ServerCommandSource>, reg
                             previous.append(enchantment)
                         }
                     }
-                    val finalPrevious = previous
-                    context.source.sendFeedback({ finalPrevious }, false)
+
+                    if (previous != null) {
+                        context.source.sendFeedback({ previous }, false)
+                    }
                     Command.SINGLE_SUCCESS
                 })
         )